api/pyproject.toml
[tool.poetry]
name = "sxt_2023"
version = "0.1.0"
description = ""
authors = ["juanmarcos <juan.cabezas@slash-digital.io>"]
license = "Proprietary"

packages = [
    {include = "sxt_2023"},
]

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.3"
gunicorn = "^21.0.0"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
django-wailer = "^1.0.0b4"
psycopg2-binary = "^2.9.9"
django-postgres-extra = "^2.0.8"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
ipython = "*"

[tool.isort]
profile = "black"
known_first_party = "sxt_2023"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
